build: add all example environments and fix cross-platform compile issues
Enable dev_arduino_* envs for all buildable examples in platformio.ini.

Compile fixes (all 10 dev envs now build cleanly):
- src/TerminalTokenizer.cpp: replace aggregate init with explicit ctor
  call for Arduino C++14 compatibility
- src/hal/common/DefaultStorageMedia.h: prevent ESP-IDF SDMMC header
  being pulled in on Arduino (guard to ESP_PLATFORM && !ARDUINO)
- src/hal/espidf/ESPIDFSDMMCStorageMedia.h: same guard, restricts class
  definition to ESP-IDF only
- src/hal/espidf/ESPNetworkInterface.h: skip ICMP ping on Arduino builds
- platformio.ini: add FS library include path for Arduino example base;
  keep deep+ LDF and pre:link_examples.py for all Arduino example envs
- examples/*.ino: namespace qualifications, uptime type casts, include
  guards for Arduino/ESP-IDF headers across Basic, Custom, DualNetwork,
  Network, Ownership, SelectiveRegistration, SimpleCustomCommand

+++ examples/CustomCommand/CustomCommand.ino
@@ -66,12 +66,12 @@ public:
 
     ETString result = "System uptime: ";
     if (days > 0)
-      result += toETString(days) + "d ";
+      result += toETString(static_cast<size_t>(days)) + "d ";
     if (hours > 0)
-      result += toETString(hours) + "h ";
+      result += toETString(static_cast<size_t>(hours)) + "h ";
     if (minutes > 0)
-      result += toETString(minutes) + "m ";
-    result += toETString(seconds) + "s";
+      result += toETString(static_cast<size_t>(minutes)) + "m ";
+    result += toETString(static_cast<size_t>(seconds)) + "s";
 
     return result;
   }
